refactor(requirements): share a single PackageJson type
Five requirements each declared their own local `PackageJson` shape. Hoist one
canonical, index-signature-backed type into `workspaces.ts` (next to
`readPackageJson`, which now defaults its generic to it) and reuse it from
`connectClosure`, `requireConnectPublicDependencies`, `requireForbiddenDeps`,
`requireUnifiedDependencyVersions` and `requirePackageJsonScripts`.

Type-only change; no runtime behavior is affected. `requirePublishConfig` keeps
its specialized type (it models the nested `publishConfig.exports` shape).
