+++ packages/requirements/src/workspaces.ts
@@ -68,5 +68,27 @@ export const getWorkspaceDirectoryMap = (repoRoot: string): ReadonlyMap<string, 
     return map;
 };
 
-export const readPackageJson = <T>(workspaceDir: string): T =>
+/**
+ * Structural view of a `package.json`. Only the fields the requirements actually
+ * read are typed explicitly; the trailing index signature keeps arbitrary
+ * metadata (`repository`, `bugs`, `author`, …) accessible as `unknown` without
+ * every consumer redeclaring its own subset.
+ */
+export type PackageJson = {
+    readonly name?: string;
+    readonly version?: string;
+    readonly private?: boolean;
+    readonly type?: string;
+    readonly scripts?: Record<string, string | undefined>;
+    readonly workspaces?: { readonly packages?: ReadonlyArray<string> } | ReadonlyArray<string>;
+    readonly resolutions?: Record<string, string>;
+    readonly dependencies?: Record<string, string>;
+    readonly devDependencies?: Record<string, string>;
+    readonly optionalDependencies?: Record<string, string>;
+    readonly peerDependencies?: Record<string, string>;
+    readonly peerDependenciesMeta?: Record<string, { readonly optional?: boolean }>;
+    readonly [field: string]: unknown;
+};
+
+export const readPackageJson = <T = PackageJson>(workspaceDir: string): T =>
     JSON.parse(readFileSync(join(workspaceDir, 'package.json'), 'utf-8')) as T;
